src/backend/requirements.txt
backports.functools-lru-cache==1.4
backports.shutil-get-terminal-size==1.0.0
certifi==2017.11.5
chardet==3.0.4
click==6.7
coreapi==2.3.3
coreschema==0.0.4
cycler==0.10.0
decorator==4.1.2
dj-database-url==0.4.2
dj-static==0.0.6
Django==1.11.7
django-appconf==1.0.2
django-compressor==2.2
django-filter==1.1
django-rest-framework-mongoengine==3.3.1
djangorestframework==3.7.7
drf-nested-routers==0.90.0
drf-yasg==1.1.3
enum34==1.1.6
flex==6.11.1
funcsigs==1.0.2
functools32==3.2.3.post2
future==0.16.0
gunicorn==19.1.1
idna==2.6
inflection==0.3.1
ipython==5.5.0
ipython-genutils==0.2.0
itypes==1.1.0
Jinja2==2.10
jsonpointer==1.14
jsonschema==2.6.0
jupyter-core==4.4.0
MarkupSafe==1.0
matplotlib==2.1.1
mock==2.0.0
mongoengine==0.15.0
nbformat==4.4.0
networkx==2.0
numpy==1.13.3
openapi-codec==1.3.2
pandas==0.22.0
pathlib2==2.3.0
pbr==3.1.1
pexpect==4.3.1
pickleshare==0.7.4
plotly==2.2.3
prompt-toolkit==1.0.15
ptyprocess==0.5.2
Pygments==2.2.0
pymongo==3.6.0
pyparsing==2.2.0
python-dateutil==2.6.1
pytz==2017.3
PyYAML==3.12
rcssmin==1.0.6
requests==2.18.4
rfc3987==1.3.7
rjsmin==1.0.12
ruamel.ordereddict==0.4.13
ruamel.yaml==0.15.35
scandir==1.6
simplegeneric==0.8.1
six==1.11.0
static3==0.7.0
strict-rfc3339==0.7
subprocess32==3.2.7
swagger-spec-validator==2.1.0
traitlets==4.3.2
typing==3.6.2
uritemplate==3.0.0
urllib3==1.22
validate-email==1.3
wcwidth==0.1.7
